pyproject.toml
[project]
name = "medicalreportanalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.0.10",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
